# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.5.0
  description: |
    Valo multi-tenant tender platform API.

    Except for health and organisation discovery/bootstrap operations, protected
    tenant operations resolve an organisation access context on the server. The
    `X-Valo-Organisation-Id` header selects a tenant; it is mandatory when a
    caller has more than one active direct membership, when accessing a client
    through an approved partner relationship, and whenever break-glass access is
    used. The header is only a selector: the server independently validates the
    authenticated user's active membership, role grants, access windows, partner
    relationship, commercial activation, and break-glass approval.
servers:
  - url: /api
    description: Base API path
...
security:
  - clerkBearer: []
paths:
...
  /partner-relationships:
    parameters:
      - $ref: "#/components/parameters/OrganisationContextHeader"
    get:
      operationId: listPartnerRelationships
      tags: [Partner Access]
      summary: List partner relationships for the selected organisation
      description: |
        Requires `partner_relationship:read`. Results are limited to rows where
        the selected organisation is either the consultancy partner or client.
      responses:
        "200":
          description: Partner relationships visible to the selected organisation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerRelationship"
        "400":
          $ref: "#/components/responses/TenantSelectionRequired"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createPartnerRelationship
      tags: [Partner Access]
      summary: Request a consultancy partner relationship
      description: |
        Requires `partner_relationship:manage` from a direct membership in a
        consultancy partner organisation. The target must be an active client
        organisation. Creation does not grant client-tenant access; the client
        must separately approve the pending relationship.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerRelationshipCreate"
      responses:
        "201":
          description: Pending relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerRelationship"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /partner-relationships/{id}/approve:
    post:
      operationId: approvePartnerRelationship
      tags: [Partner Access]
      summary: Approve a pending partner relationship
      description: |
        Requires `partner_relationship:manage` from a direct membership in the
        client organisation named by the relationship. The client tenant must
        have `partner_edition` commercially activated. The relationship version
        supplied in `If-Match` prevents stale approval.
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/OrganisationContextHeader"
        - $ref: "#/components/parameters/IfMatchVersion"
      responses:
        "200":
          description: Active partner relationship
          headers:
            ETag:
              $ref: "#/components/headers/VersionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerRelationship"
        "400":
          $ref: "#/components/responses/TenantSelectionRequired"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token. Authentication never implies access to a tenant.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    OrganisationContextHeader:
      name: X-Valo-Organisation-Id
      in: header
      required: false
      description: |
        Selects the organisation access context. It is required when the caller
        has multiple active direct memberships and always required for a
        partner-managed client tenant or break-glass access. It may be omitted
        only when the server can unambiguously select the caller's sole active
        direct membership. This value never grants access by itself: the server
        revalidates membership, role grants, access windows, relationship and
        commercial state for every request.
      schema:
        type: string
        format: uuid
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: Current positive integer resource version, optionally quoted or weakly prefixed.
      schema:
        type: string
        pattern: '^(?:W/)?"?[1-9][0-9]*"?$'
      example: '"3"'
...
  headers:
    VersionETag:
      description: Quoted optimistic-concurrency version for the returned resource.
      schema:
        type: string
      example: '"4"'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Account, tenant context, role or permission does not authorise the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TenantSelectionRequired:
      description: Tenant selection is missing or incompatible with another context header
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PreconditionRequired:
      description: A current resource version is required in If-Match
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalServerError:
      description: Authentication or tenant-context resolution failed unexpectedly
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PartnerRelationship:
      type: object
      additionalProperties: false
      required:
        - id
        - partnerOrganisationId
        - clientOrganisationId
        - status
        - clientOwnershipRule
        - qaResponsibility
        - coSigningRequired
        - accessStartsAt
        - accessExpiresAt
        - approvedByMembershipId
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        partnerOrganisationId:
          type: string
          format: uuid
        clientOrganisationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, active]
        clientOwnershipRule:
          type: string
          enum: [client_retained]
        qaResponsibility:
          type: string
          maxLength: 500
          nullable: true
        coSigningRequired:
          type: boolean
        accessStartsAt:
          type: string
          format: date-time
          nullable: true
        accessExpiresAt:
          type: string
          format: date-time
          nullable: true
        approvedByMembershipId:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PartnerRelationshipCreate:
      type: object
      additionalProperties: false
      required: [clientOrganisationId]
      properties:
        clientOrganisationId:
          type: string
          format: uuid
        qaResponsibility:
          type: string
          maxLength: 500
          nullable: true
        coSigningRequired:
          type: boolean
          default: false
        accessExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: Must be in the future when supplied.
...